--- keyboards/crkbd/keymaps/bbaserdem-arm/config.h
@@ -1,53 +1,0 @@
-#pragma once
-#include "config_common.h"
-#undef PRODUCT
-#undef MANUFACTURER
-#define MANUFACTURER Waffles
-#define PRODUCT Protorne Keyboard
-#undef USE_I2C
-#undef SSD1306OLED
-#define EE_HANDS
-#define USE_SERIAL_PD2
-#undef RGBLED_NUM
-#undef RGB_DI_PIN
-#ifdef RGB_MATRIX_ENABLE
-#define RGB_MATRIX_MAXIMUM_BRIGHTNESS 150
-#define RGB_MATRIX_SPLIT { 27, 27 }
-// #define RGB_MATRIX_KEYPRESSES
-// #define RGB_MATRIX_FRAMEBUFFER_EFFECTS
-// #define RGB_DISABLE_WHEN_USB_SUSPENDED true
-// #define SPLIT_TRANSPORT_MIRROR
-#endif
-#ifdef RGBLIGHT_ENABLE
-#define RGBLIGHT_ANIMATIONS
-#undef RGBLED_SPLIT
-#define RGBLED_SPLIT { 27, 27 }
-#define RGBLIGHT_LIMIT_VAL 175
-#endif
-#define RGBLED_NUM 54
-#define RGB_DI_PIN B5
-#define WS2812_PWM_DRIVER PWMD3
-#define WS2812_PWM_CHANNEL 3
-#define WS2812_PWM_PAL_MODE 2
-#define WS2812_DMA_STREAM STM32_DMA1_STREAM3
-#define WS2812_DMA_CHANNEL 3
-#undef SOFT_SERIAL_PIN
-#define SOFT_SERIAL_PIN D3
-#define SERIAL_USART_SPEED 921600
-#define SERIAL_USART_DRIVER SD1
-#define SERIAL_USART_TX_PAL_MODE 7
-#ifdef ENCODER_ENABLE
-#define ENCODERS_PAD_A { B2 }
-#define ENCODERS_PAD_B { B6 }
-#define ENCODERS_PAD_A_RIGHT { B2 }
-#define ENCODERS_PAD_B_RIGHT { B6 }
-#define ENCODER_RESOLUTIONS { 4, 2 }
-#endif
-#ifdef PIMORONI_TRACKBALL
-#define PIMORONI_TRACKBALL_ROTATE
-#define PIMORONI_TRACKBALL_INVERT_Y 
-#endif
-#ifdef AUDIO_ENABLE
-#define AUDIO_PIN A5
-#define STARTUP_SONG SONG(ODE_TO_JOY)
-#endif
