--- keyboards/crkbd/keymaps/bbaserdem-arm/halconf.h
@@ -1,5 +1,0 @@
-#pragma once
-// #define HAL_USE_DAC TRUE
-#define HAL_USE_PWM TRUE
-#define HAL_USE_SERIAL TRUE
-#include_next <halconf.h>
