--- keyboards/crkbd/keymaps/bbaserdem-arm/rules.mk
@@ -1,22 +1,0 @@
-USER_NAME := bbaserdem
-MCU = STM32F303
-CTPC = yes
-DEBUG_MATRIX_SCAN_RATE_ENABLE = no
-SERIAL_DRIVER = usart
-WS2812_DRIVER = pwm
-VIA_ENABLE = yes
-OLED_ENABLE = yes
-WPM_ENABLE = yes
-AUDIO_ENABLE = yes #breaks reset keycode?
-AUDIO_DRIVER = dac_basic
-EXTRAKEY_ENABLE = yes
-COMMAND_ENABLE = no
-CONSOLE_ENABLE = no
-BOOTMAGIC_ENABLE = lite
-SPLIT_KEYBOARD = yes
-RGBLIGHT_ENABLE = no
-RGB_MATRIX_ENABLE = yes
-LTO_ENABLE = yes
-NKRO_ENABLE = yes
-ENCODER_ENABLE = yes
-PIMORONI_TRACKBALL = no
