+++ layouts/community/split_3x6_3/bbaserdem/halconf.h
