+++ layouts/community/split_3x6_3/bbaserdem/keymap.c
@@ -1,3 +1,16 @@
+/*
+Copyright 2019 Batuhan Başerdem <baserdem.batuhan@gmail.com> @bbaserdem
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 2 of the License, or
+(at your option) any later version.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
 #include "bbaserdem.h"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
@@ -63,17 +76,21 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     )
 };
 
-
-
 #ifdef AUDIO_ENABLE
-void keyboard_pre_init_kb(void) { //thank you to @sigprof for this
+// crkbd
+#if defined(KEYBOARD_crkbd_rev1)
+// Audio requires some fixes to work on @waffle's PCB
+// ! Copy pasted from ItsWaffle's waffle fork
+void keyboard_pre_init_keymap(void) {
+    //thank you to @sigprof for this
     // Set audio pins to analog mode
     palSetLineMode(A5, PAL_MODE_INPUT_ANALOG);
     palSetLineMode(B1, PAL_MODE_INPUT_ANALOG);
 }
 
-void keyboard_post_init_kb(void) {
+void keyboard_post_init_keymap(void) {
     // Enable OPAMP1 as A5 → B1 follower
     OPAMP3->CSR = OPAMP3_CSR_VMSEL_1 | OPAMP3_CSR_VMSEL_0 | OPAMP3_CSR_VPSEL_0 | OPAMP3_CSR_OPAMP3EN;
 }
-#endif //audio
+#endif // End of differentiations
+#endif // End of audio specific stuff
